chore: move most website-related modules to 'website'
It's getting quite crowded with `.py` files in the root of the
repository. Move most of the modules that have something to do
with the website to a submodule called `website`.
