+++ app.py
@@ -8,7 +8,6 @@ import datetime
 import collections
 import hedy
 import json
-import jsonbin
 import logging
 import os
 from os import path
@@ -19,7 +18,7 @@ from ruamel import yaml
 from flask_commonmark import Commonmark
 from werkzeug.urls import url_encode
 from config import config
-from auth import auth_templates, current_user, requires_login, is_admin, is_teacher
+from website.auth import auth_templates, current_user, requires_login, is_admin, is_teacher
 from utils import db_get, db_get_many, db_create, db_update, timems, type_check, object_check, db_del, load_yaml, load_yaml_rt, dump_yaml_rt, version
 import utils
 
@@ -31,11 +30,7 @@ from flask_compress import Compress
 # Hedy-specific modules
 import courses
 import hedyweb
-import translating
-import querylog
-import aws_helpers
-import ab_proxying
-import cdn
+from website import querylog, aws_helpers, jsonbin, translating, ab_proxying, cdn
 
 # Set the current directory to the root Hedy folder
 os.chdir(os.path.join (os.getcwd (), __file__.replace (os.path.basename (__file__), '')))
@@ -830,7 +825,7 @@ def update_yaml():
 
 # *** AUTH ***
 
-import auth
+from website import auth
 auth.routes (app, requested_lang)
 
 # *** START SERVER ***
